Fix SimplifyFormRenderingRector replacing calls other than createView()
The SimplifyFormRenderingRector was replacing calls to methods other than ->createView(), which is no longer required in Symfony 6.2, e.g. calls to ->getData().

These are not deprecated and there's no reason to remove them, and in fact the blanket rule of removing any method calls of FormInterface parameters of the container $this->render() function is too wide and could lead to breakage.